# Improve performance of method haveOverlap